Relacion de tabla areas y form para eliminar

+++ pages/Admin/area/delete_area.php
@@ -3,6 +3,15 @@
     session_start();
     if (isset($_GET['id'])){
         $id = $_GET['id'];
+        $query = "SELECT * FROM areas WHERE id = $id";
+        $result = mysqli_query($conn, $query);
+        $row = mysqli_fetch_array($result);
+        $name_area = $row['name'];
+        $query = "DELETE FROM forms WHERE name_area= '$name_area'";
+        $result = mysqli_query($conn, $query);
+        if(!$result){
+            die("Failed");
+        }
         $query = "DELETE FROM areas WHERE id= $id";
         $result = mysqli_query($conn, $query);
         if(!$result){
